directory struct change:improving readability for streams.py
