+++ tests/unittests/test_client.py
@@ -21,18 +21,18 @@ class Mockresponse:
 
 
 def mocked_failed_429_request(*args, **kwargs):
-    return Mockresponse('', 429, headers={}, raise_error=True)
+    return Mockresponse("", 429, headers={}, raise_error=True)
 
 
-@patch('requests.Session.request', side_effect=mocked_failed_429_request)
+@patch("requests.Session.request", side_effect=mocked_failed_429_request)
 def test_too_many_requests_429_error(mocked):
     client = Client(api_token="test")
 
     try:
         # Verifying if the custom exception 'DixaClient429Error' is raised on receiving status code 429
-        response = client.get("https://test.com", "/test")
-    except DixaClient429Error as e:
+        _ = client.get("https://test.com", "/test")
+    except DixaClient429Error as api_error:
         expected_error_message = "API limit has been reached"
 
         # Verifying the message formed for the custom exception
-        assert str(e) == expected_error_message
+        assert str(api_error) == expected_error_message
